Refactor tests of POSCAR imports
