+++ src/molara/structure/io/importer_crystal.py
@@ -61,6 +61,9 @@ class PymatgenImporter(Importer):
     def load(self) -> Crystals:
         """Import a file and returns the Crystal."""
         try:
+            # from monty.io import zopen
+            # with zopen(self.path, "rt", errors="replace") as f:
+            #     contents = f.read()
             from pymatgen.core import Structure
 
             structure = Structure.from_file(self.path)
@@ -97,13 +100,7 @@ class PoscarImporter(Importer):
         """
         if use_pymatgen:
             try:
-                from monty.io import zopen
-                from pymatgen.core import Structure as PymatgenStructure
-
-                with zopen(self.path, "rt", errors="replace") as f:
-                    contents = f.read()
-                structure = PymatgenStructure.from_str(contents, fmt="poscar")
-                crystal = Crystal.from_pymatgen(structure, supercell_dims=[1, 1, 1])
+                crystal = PymatgenImporter(self.path).load().get_current_mol()
             except ImportError:
                 warnings.warn("pymatgen is not installed, using internal parser", stacklevel=2)
                 crystal = self.parse_poscar()
